requirements.txt
pillow
pytesseract
selenium==4.26.1
mysql-connector-python
pyautogui; sys_platform == "linux"